pdf_helper: rename to pdf_xml_attachment
