--- pdf_helper/readme/DESCRIPTION.md
@@ -1,1 +1,0 @@
-Technical module to share PDF utils.
