+++ pdf_xml_attachment/README.rst
@@ -1,6 +1,6 @@
-==========
-PDF Helper
-==========
+==================
+PDF XML attachment
+==================
 
 .. 
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -17,10 +17,10 @@ PDF Helper
     :target: http://www.gnu.org/licenses/lgpl-3.0-standalone.html
     :alt: License: LGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Freporting--engine-lightgray.png?logo=github
-    :target: https://github.com/OCA/reporting-engine/tree/18.0/pdf_helper
+    :target: https://github.com/OCA/reporting-engine/tree/18.0/pdf_xml_attachment
     :alt: OCA/reporting-engine
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/reporting-engine-18-0/reporting-engine-18-0-pdf_helper
+    :target: https://translation.odoo-community.org/projects/reporting-engine-18-0/reporting-engine-18-0-pdf_xml_attachment
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
     :target: https://runboat.odoo-community.org/builds?repo=OCA/reporting-engine&target_branch=18.0
@@ -28,7 +28,7 @@ PDF Helper
 
 |badge1| |badge2| |badge3| |badge4| |badge5|
 
-Technical module to share PDF utils.
+Technical module to share easily deal with XML attachments in PDF files.
 
 **Table of contents**
 
@@ -42,15 +42,15 @@ Inside Odoo env:
 
 ::
 
-   res = env["pdf.helper"].pdf_get_xml_files(pdf_filecontent)
+   res = env["pdf.xml.tool"].pdf_get_xml_files(pdf_filecontent)
 
-   new_pdf_filecontent = env["pdf.helper"].pdf_embed_xml(pdf_filecontent, filename, xml)
+   new_pdf_filecontent = env["pdf.xml.tool"].pdf_embed_xml(pdf_filecontent, filename, xml)
 
 Outside Odoo env:
 
 ::
 
-   from odoo.addons.pdf_helper.utils import PDFParser
+   from odoo.addons.pdf_xml_attachment.utils import PDFParser
    [...]
    res = PDFParser(pdf_filecontent).get_xml_files()
 
@@ -60,7 +60,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/reporting-engine/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/reporting-engine/issues/new?body=module:%20pdf_helper%0Aversion:%2018.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/reporting-engine/issues/new?body=module:%20pdf_xml_attachment%0Aversion:%2018.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -103,6 +103,6 @@ Current `maintainers <https://odoo-community.org/page/maintainer-role>`__:
 
 |maintainer-simahawk| |maintainer-alexis-via| 
 
-This module is part of the `OCA/reporting-engine <https://github.com/OCA/reporting-engine/tree/18.0/pdf_helper>`_ project on GitHub.
+This module is part of the `OCA/reporting-engine <https://github.com/OCA/reporting-engine/tree/18.0/pdf_xml_attachment>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
