+++ pdf_xml_attachment/__manifest__.py
@@ -2,13 +2,12 @@
 # @author: Simone Orsi <simahawk@gmail.com>
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
-# TODO: move it to a simple python package under OCA umbrella?
 {
-    "name": "PDF Helper",
+    "name": "PDF XML attachment",
     "version": "18.0.1.0.0",
     "category": "Tools",
     "license": "LGPL-3",
-    "summary": "Provides helpers to work w/ PDFs",
+    "summary": "Provides helpers to work w/ PDFs and XML attachments",
     "author": "Camptocamp, Odoo Community Association (OCA)",
     "maintainers": ["simahawk", "alexis-via"],
     "website": "https://github.com/OCA/reporting-engine",
