+++ pdf_xml_attachment/i18n/it.po
@@ -1,22 +1,22 @@
 # Translation of Odoo Server.
 # This file contains the translation of the following modules:
-# 	* pdf_helper
+# 	* pdf_xml_attachment
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
+"Project-Id-Version: Odoo Server 17.0\n"
 "Report-Msgid-Bugs-To: \n"
-"PO-Revision-Date: 2023-08-07 11:09+0000\n"
-"Last-Translator: Ivorra78 <informatica@totmaterial.es>\n"
+"PO-Revision-Date: 2024-02-04 23:45+0000\n"
+"Last-Translator: mymage <stefano.consolaro@mymage.it>\n"
 "Language-Team: none\n"
-"Language: es\n"
+"Language: it\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: \n"
 "Plural-Forms: nplurals=2; plural=n != 1;\n"
 "X-Generator: Weblate 4.17\n"
 
-#. module: pdf_helper
-#: model:ir.model,name:pdf_helper.model_pdf_helper
-msgid "PDF Helper"
-msgstr "Ayudante de PDF"
+#. module: pdf_xml_attachment
+#: model:ir.model,name:pdf_xml_attachment.model_pdf_xml_attachment
+msgid "PDF XML attachment"
+msgstr "Aiuto PDF"
