+++ pdf_xml_attachment/readme/DESCRIPTION.md
@@ -1,0 +1,1 @@
+Technical module to share easily deal with XML attachments in PDF files.
