+++ pdf_xml_attachment/tests/test_helper.py
@@ -8,7 +8,7 @@ from lxml import etree
 from odoo.tests.case import TestCase
 from odoo.tests.common import TransactionCase
 
-from odoo.addons.pdf_helper.utils import PDFParser
+from odoo.addons.pdf_xml_attachment.utils import PDFParser
 
 
 def read_test_file(filename, mode="r"):
@@ -29,16 +29,16 @@ class TestPDFHelperUtils(TestCase):
 class TestPDFHelper(TransactionCase):
     def test_get_xml(self):
         pdf_content = read_test_file("pdf_with_xml_test.pdf", mode="rb")
-        res = self.env["pdf.helper"].pdf_get_xml_files(pdf_content)
+        res = self.env["pdf.xml.tool"].pdf_get_xml_files(pdf_content)
         fname, xml_root = tuple(res.items())[0]
         self.assertEqual(fname, "factur-x.xml")
         self.assertTrue(isinstance(xml_root, etree._Element))
 
     def test_get_xml_fail(self):
         with self.assertLogs(
-            "odoo.addons.pdf_helper.models.helper", level="ERROR"
+            "odoo.addons.pdf_xml_attachment.models.helper", level="ERROR"
         ) as log_catcher:
-            self.env["pdf.helper"].pdf_get_xml_files(b"")
+            self.env["pdf.xml.tool"].pdf_get_xml_files(b"")
             self.assertIn(
                 "PDF file parsing failed: Cannot read an empty file",
                 log_catcher.output[0],
@@ -48,10 +48,10 @@ class TestPDFHelper(TransactionCase):
         pdf_content = read_test_file("pdf_with_xml_test.pdf", mode="rb")
         filename = "test"
         xml = b"<root>test</root>"
-        newpdf_content = self.env["pdf.helper"].pdf_embed_xml(
+        newpdf_content = self.env["pdf.xml.tool"].pdf_embed_xml(
             pdf_content, filename, xml
         )
-        attachments = self.env["pdf.helper"].pdf_get_xml_files(newpdf_content)
+        attachments = self.env["pdf.xml.tool"].pdf_get_xml_files(newpdf_content)
         self.assertTrue(filename in attachments)
         etree_content = attachments[filename]
         self.assertEqual(xml, etree.tostring(etree_content))
